Summary of changes from v2.4.19-pre5 to v2.4.19-pre6 ============================================
<vojtech@suse.cz> (02/04/01 1.311)
	[PATCH] Fixes for non-legacy gameports.

<manfred@colorfullife.com> (02/04/01 1.312)
	[PATCH] block/IDE/interrupt lockup fix

<martin@meltin.net> (02/04/01 1.313)
	[PATCH] Patch to pull NFS server address off root_server_path

<akpm@zip.com.au> (02/04/01 1.314)
	[PATCH] Let all hptraid ioctls to throught the block layer

<viro@math.psu.edu> (02/04/01 1.315)
	Cleans do_remount_sb() up and docbookifies it.

<viro@math.psu.edu> (02/04/01 1.316)
	Slightly cleans up the handling of anon. devices,

<viro@math.psu.edu> (02/04/01 1.317)
	Obvious cleanups in get_sb_bdev().

<viro@math.psu.edu> (02/04/01 1.318)
	* capability check moved from do_kern_mount() into do_add_mount().

<viro@math.psu.edu> (02/04/01 1.319)
	rootfs made an alias of ramfs.

<viro@math.psu.edu> (02/04/01 1.320)
	Removes lock_super()/unlock_super() from callers of ->read_super().

<viro@math.psu.edu> (02/04/01 1.321)
	mount_sem turned into rwsem.  The only reader is handling of

<viro@math.psu.edu> (02/04/01 1.322)
	turns (mount_sem,vfsmntlist,root_vfsmnt) into per-process object

<viro@math.psu.edu> (02/04/01 1.323)
	makes /proc/mounts a symlink to /proc/<pid>/mounts.

<viro@math.psu.edu> (02/04/01 1.324)
	kills set_devname(), makes "name" an argument of alloc_vfsmnt().

<bcrl@redhat.com> (02/04/01 1.325)
	The patch below fixes a problem whereby a vma which has its vm_start

<arjanv@redhat.com> (02/04/02 1.326)
	[PATCH] dma_addr_t vs highmem

<greg@kroah.com> (02/04/02 1.300.3.1)
	USB visor driver

<greg@kroah.com> (02/04/02 1.300.3.2)
	USB HID driver fixes

<bhards@bigpond.net.au> (02/04/02 1.300.3.3)
	USB CDCEther driver update

<petkan@mastika.lnxw.com> (02/04/02 1.300.3.4)
	USB

<david-b@pacbell.net> (02/04/02 1.300.3.5)
	USB hcd core

<david-b@pacbell.net> (02/04/02 1.300.3.6)
	USB ohci driver fixes

<petkan@mastika.lnxw.com> (02/04/02 1.300.3.7)
	USB pegasus driver

<suse.cz@mastika.lnxw.com> (02/04/02 1.300.3.8)
	USB

<johannes@erdfelt.com> (02/04/02 1.300.3.9)
	USB uhci

<greg@kroah.com> (02/04/02 1.300.3.10)
	USB

<stelian@popies.net> (02/04/03 1.327)
	Enable more accurate events on Vaio laptops without a jogdial (FX series)

<kaos@ocs.com.au> (02/04/03 1.326.1.1)
	[PATCH] 2.4.19-pre5 Makefile standardization

<jaharkes@cs.cmu.edu> (02/04/03 1.326.1.2)
	[PATCH] 2.4.19-pre5 Coda update

<marcelo@plucky.distro.conectiva> (02/04/03 1.326.1.3)
	Removed the EXPERIMENTAL mark of fs/Config.in ext3 entry.

<marcelo@plucky.distro.conectiva> (02/04/03 1.329)
	Cosmetic fix to avoid the agpgart detection

<rml@tech9.net> (02/04/03 1.330)
	[PATCH] 2.4: BUG_ON (1/2)

<rml@tech9.net> (02/04/03 1.331)
	[PATCH] 2.4: BUG_ON (2/2)

<marcelo@plucky.distro.conectiva> (02/04/03 1.332)
	Re add MWAVE Config.in entry

<rusty@rustcorp.com.au> (02/04/03 1.333)
	[PATCH] TRIVIAL 2.4.19-pre5: PPC leapyear fix

<marcelo@plucky.distro.conectiva> (02/04/03 1.334)
	Add ALi 1644 support to AGP

<bcrl@redhat.com> (02/04/04 1.335)
	[PATCH] Update mmap patch

<sam@minnie.(none)> (02/04/04 1.326.2.1)
	tlan.c:

<ioshadij@hotmail.com> (02/04/04 1.336)
	[PATCH] kjournald locking fix

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.337)
	[PATCH] PATCH: regenerated against new tree - Configs

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.338)
	[PATCH] PATCH: make indydog use long for bitops

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.339)
	[PATCH] PATCH: wdt285 error returns

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.340)
	[PATCH] PATCH: silly doc fix

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.341)
	[PATCH] PATCH: also add bridge resources to resource tree

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.342)
	[PATCH] PATCH: returns on error fixes for sound

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.343)
	[PATCH] PATCH: make the mad16 use the newer input/gameport api right

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.344)
	[PATCH] PATCH: yet more sound error returns

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.345)
	[PATCH] PATCH: and one or two more for luck 8)

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.346)
	[PATCH] PATCH: only flush block on the last close

<alan@lxorguk.ukuu.org.uk> (02/04/04 1.347)
	[PATCH] PATCH: missing neomagic include bits

<hch@infradead.org> (02/04/04 1.348)
	[PATCH] SARD

<viro@math.psu.edu> (02/04/04 1.349)
	[PATCH] IS_DEADDIR() checks

<rusty@rustcorp.com.au> (02/04/04 1.350)
	[PATCH] TRIVIAL 2.4.19-pre5: documentation fix

<trini@kernel.crashing.org> (02/04/04 1.351)
	[PATCH] Don't always ask about Intel or AMD RNGs

<rusty@rustcorp.com.au> (02/04/05 1.353)
	[PATCH] TRIVIAL 2.4.19-pre5: fcntl (F_DUPFD) return

<neilb@cse.unsw.edu.au> (02/04/05 1.354)
	[PATCH] PATCH 1 of 16 - Fix problems with raid1 resync code.

<neilb@cse.unsw.edu.au> (02/04/05 1.355)
	[PATCH] PATCH 2 of 16 - Flush out final sync requests on an idle system.

<neilb@cse.unsw.edu.au> (02/04/05 1.356)
	[PATCH] PATCH 3 of 16 - Remove exp_find, it is never used

<neilb@cse.unsw.edu.au> (02/04/05 1.357)
	[PATCH] PATCH 4 of 16 - read_lock the export table when lockd calls

<neilb@cse.unsw.edu.au> (02/04/05 1.358)
	[PATCH] PATCH 5 of 16 - Fix possible leak of mnt/dentry references.

<neilb@cse.unsw.edu.au> (02/04/05 1.359)
	[PATCH] PATCH 6 of 16 - Use MKDEV for making device number from components

<neilb@cse.unsw.edu.au> (02/04/05 1.360)
	[PATCH] PATCH 7 of 16 - Central updating of fh_stale statistics.

<neilb@cse.unsw.edu.au> (02/04/05 1.361)
	[PATCH] PATCH 8 of 16 - Get nfsd_setattr to not put too much weight on

<neilb@cse.unsw.edu.au> (02/04/05 1.362)
	[PATCH] PATCH 9 of 16 - Tidy up some vfs calls in nfsd

<neilb@cse.unsw.edu.au> (02/04/05 1.363)
	[PATCH] PATCH 12 of 16 - Stop fat_fh_to_dentry returning NULL + set

<neilb@cse.unsw.edu.au> (02/04/05 1.364)
	[PATCH] PATCH 10 of 16 - Cleanup the syscall interface to nfsd

<neilb@cse.unsw.edu.au> (02/04/05 1.365)
	[PATCH] PATCH 13 of 16 - Tidy up exp_get code

<neilb@cse.unsw.edu.au> (02/04/05 1.366)
	[PATCH] PATCH 14 of 16 - Change exports hash lists to list.h lists

<neilb@cse.unsw.edu.au> (02/04/05 1.367)
	[PATCH] PATCH 15 of 16 - Link exports for a given client together

<neilb@cse.unsw.edu.au> (02/04/05 1.368)
	[PATCH] PATCH 16 of 16 - Change /proc/fs/nfs/exports to use seq_file

<viro@math.psu.edu> (02/04/05 1.369)
	kills floppy_eject(), replacing it with normal open()/ioctl()/close().

<viro@math.psu.edu> (02/04/05 1.370)
	moves initrd-related options (rd_doload, etc.) into do_mounts.c

<viro@math.psu.edu> (02/04/05 1.371)
	switches wait_for_keypress() to normal syscalls.

<viro@math.psu.edu> (02/04/05 1.372)
	moves devfs_make_root() to do_mounts.c and cleans it up.

<Lionel.Bouton@inet6.fr> (02/04/05 1.373)
	[PATCH] Against 2.4.19-pre5, Bugfixes

<kaos@ocs.com.au> (02/04/05 1.374)
	[PATCH] 2.4.19-pre5 prevent user space includes

<marcelo@plucky.distro.conectiva> (02/04/05 1.375)
	Changed EXTRAVERSION to pre6

<neilb@cse.unsw.edu.au> (02/04/05 1.376)
	[PATCH] Re: PATCH 11 of 16 - Tidyup init/exit fof nfsd module

+++ drivers/usb/hcd/ehci-hcd.c
@@ -747,8 +747,8 @@ static int __init init (void)
 {
 	dbg (DRIVER_INFO);
 	dbg ("block sizes: qh %d qtd %d itd %d sitd %d",
-		sizeof (struct ehci_qh), sizeof (struct ehci_qtd),
-		sizeof (struct ehci_itd), sizeof (struct ehci_sitd));
+		(int) sizeof (struct ehci_qh), (int) sizeof (struct ehci_qtd),
+		(int) sizeof (struct ehci_itd), (int) sizeof (struct ehci_sitd));
 
 	return pci_module_init (&ehci_pci_driver);
 }
